internal: add util mod for test
