+++ crates/moon/tests/mod.rs
@@ -18,13 +18,10 @@
 
 mod cmd_test;
 mod test_cases;
+mod util;
 
-use expect_test::Expect;
 use std::path::{Path, PathBuf};
-
-fn check(actual: &str, expect: Expect) {
-    expect.assert_eq(actual)
-}
+use util::*;
 
 struct TestDir {
     // tempfile::TempDir has a drop implementation that will remove the directory
@@ -58,10 +55,6 @@ impl AsRef<Path> for TestDir {
     fn as_ref(&self) -> &Path {
         self.path.path()
     }
-}
-
-pub fn moon_bin() -> PathBuf {
-    snapbox::cmd::cargo_bin("moon")
 }
 
 #[track_caller]
@@ -100,6 +93,23 @@ fn get_stderr_without_replace(
     s
 }
 
+fn get_err_stdout_without_replace(
+    dir: &impl AsRef<std::path::Path>,
+    args: impl IntoIterator<Item = impl AsRef<std::ffi::OsStr>>,
+) -> String {
+    let out = snapbox::cmd::Command::new(moon_bin())
+        .current_dir(dir)
+        .args(args)
+        .assert()
+        .failure()
+        .get_output()
+        .stdout
+        .to_owned();
+
+    let s = std::str::from_utf8(&out).unwrap().to_string();
+    s
+}
+
 #[track_caller]
 fn get_err_stderr_without_replace(
     dir: &impl AsRef<std::path::Path>,
@@ -118,27 +128,6 @@ fn get_err_stderr_without_replace(
     s
 }
 
-pub fn replace_dir(s: &str, dir: &impl AsRef<std::path::Path>) -> String {
-    let path_str1 = dunce::canonicalize(dir)
-        .unwrap()
-        .to_str()
-        .unwrap()
-        .to_string();
-    // for something like "{...\"loc\":{\"path\":\"C:\\\\Users\\\\runneradmin\\\\AppData\\\\Local\\\\Temp\\\\.tmpP0u4VZ\\\\main\\\\main.mbt\"...\r\n" on windows
-    // https://github.com/moonbitlang/moon/actions/runs/10092428950/job/27906057649#step:13:149
-    let s = s.replace("\\\\", "\\");
-    let s = s.replace(&path_str1, "$ROOT");
-    let s = s.replace(
-        dunce::canonicalize(moonutil::moon_dir::home())
-            .unwrap()
-            .to_str()
-            .unwrap(),
-        "$MOON_HOME",
-    );
-    let s = s.replace(moon_bin().to_string_lossy().as_ref(), "moon");
-    s.replace("\r\n", "\n").replace('\\', "/")
-}
-
 #[track_caller]
 pub fn get_stdout_and_replace_dir(
     dir: &impl AsRef<std::path::Path>,
@@ -148,20 +137,28 @@ pub fn get_stdout_and_replace_dir(
     replace_dir(&s, dir)
 }
 
-pub fn get_err_stderr_and_replace_dir(
-    dir: &impl AsRef<std::path::Path>,
-    args: impl IntoIterator<Item = impl AsRef<std::ffi::OsStr>>,
-) -> String {
-    let s = get_err_stderr_without_replace(dir, args);
-    replace_dir(&s, dir)
-}
-
 #[track_caller]
 pub fn get_stderr_and_replace_dir(
     dir: &impl AsRef<std::path::Path>,
     args: impl IntoIterator<Item = impl AsRef<std::ffi::OsStr>>,
 ) -> String {
     let s = get_stderr_without_replace(dir, args);
+    replace_dir(&s, dir)
+}
+
+pub fn get_err_stdout_and_replace_dir(
+    dir: &impl AsRef<std::path::Path>,
+    args: impl IntoIterator<Item = impl AsRef<std::ffi::OsStr>>,
+) -> String {
+    let s = get_err_stdout_without_replace(dir, args);
+    replace_dir(&s, dir)
+}
+
+pub fn get_err_stderr_and_replace_dir(
+    dir: &impl AsRef<std::path::Path>,
+    args: impl IntoIterator<Item = impl AsRef<std::ffi::OsStr>>,
+) -> String {
+    let s = get_err_stderr_without_replace(dir, args);
     replace_dir(&s, dir)
 }
 
@@ -201,62 +198,4 @@ pub fn get_stderr(
     let s = s.replace("\r\n", "\n");
 
     s.replace('\\', "/")
-}
-
-pub fn copy(src: &Path, dest: &Path) -> anyhow::Result<()> {
-    if src.is_dir() {
-        if !dest.exists() {
-            std::fs::create_dir_all(dest)?;
-        }
-        for entry in walkdir::WalkDir::new(src) {
-            let entry = entry?;
-            let path = entry.path();
-            let relative_path = path.strip_prefix(src)?;
-            let dest_path = dest.join(relative_path);
-            if path.is_dir() {
-                if !dest_path.exists() {
-                    std::fs::create_dir_all(dest_path)?;
-                }
-            } else {
-                std::fs::copy(path, dest_path)?;
-            }
-        }
-    } else {
-        std::fs::copy(src, dest)?;
-    }
-    Ok(())
-}
-
-pub fn replace_crlf_to_lf(s: &str) -> String {
-    s.replace("\r\n", "\n")
-}
-
-fn get_err_stdout_without_replace(
-    dir: &impl AsRef<std::path::Path>,
-    args: impl IntoIterator<Item = impl AsRef<std::ffi::OsStr>>,
-) -> String {
-    let out = snapbox::cmd::Command::new(moon_bin())
-        .current_dir(dir)
-        .args(args)
-        .assert()
-        .failure()
-        .get_output()
-        .stdout
-        .to_owned();
-
-    let s = std::str::from_utf8(&out).unwrap().to_string();
-    s
-}
-
-pub fn get_err_stdout_and_replace_dir(
-    dir: &impl AsRef<std::path::Path>,
-    args: impl IntoIterator<Item = impl AsRef<std::ffi::OsStr>>,
-) -> String {
-    let s = get_err_stdout_without_replace(dir, args);
-    replace_dir(&s, dir)
-}
-
-#[track_caller]
-fn read(p: &Path) -> String {
-    std::fs::read_to_string(p).unwrap().replace('\r', "")
 }
